refactor: update FavTaskController, K8sNamespaceController, K8sNamespaceControllerTest
